Create README.md
The models used in the manuscript are in the "ZINB" file. Data is property of UnderSea Hunter, and referred to as "cocos_environmentals" throughout the script.  Because the data is property of UnderSea Hunter, it is not available online. 

For questions about this code please reach out to Julia.Saltzman@UNH.edu or Easton.White@UNH.edu. 


Learn more about our work here: https://quantmarineecolab.github.io/